+++ CHANGELOG.md
@@ -26,6 +26,11 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://github.com/atlassian/infrastructure/compare/release-4.8.0...master
 
+### Added
+- A way to gather thread dumps over ssh. Resolve [JPERF-405].
+
+[JPERF-405]: https://ecosystem.atlassian.net/browse/JPERF-405
+
 ## [4.8.0] - 2019-02-22
 [4.8.0]: https://github.com/atlassian/infrastructure/compare/release-4.7.0...release-4.8.0